fix no cpu build
