+++ mlx/backend/no_cpu/CMakeLists.txt
@@ -6,5 +6,6 @@ target_sources(
   ${CMAKE_CURRENT_SOURCE_DIR}/../common/common.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../common/compiled.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../common/compiled_nocpu.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/../common/reduce_utils.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/../common/slicing.cpp
 )
